Black format

+++ eust/tables/data.py
@@ -9,8 +9,8 @@ import numpy as np
 from eust.core import _download_file, conf
 
 
-_DIMENSION_NAME_RE = re.compile(r'^[a-z_0-9]+$')
-_YEAR_RE = re.compile(r'^(1|2)[0-9]{3}$')
+_DIMENSION_NAME_RE = re.compile(r"^[a-z_0-9]+$")
+_YEAR_RE = re.compile(r"^(1|2)[0-9]{3}$")
 
 
 def _is_valid_dimension_name(s: str) -> bool:
@@ -18,11 +18,13 @@ def _is_valid_dimension_name(s: str) -> bool:
 
 
 def _split_values_flags(series: pd.Series) -> pd.DataFrame:
-    split = series.str.split(' ')
-    df = pd.DataFrame({
-        'value': split.apply(lambda l: l[0] if l else None),
-        'flag': split.apply(lambda l: l[1] if l and len(l) > 1 else None)
-        })
+    split = series.str.split(" ")
+    df = pd.DataFrame(
+        {
+            "value": split.apply(lambda l: l[0] if l else None),
+            "flag": split.apply(lambda l: l[1] if l and len(l) > 1 else None),
+        }
+    )
     return df
 
 
@@ -34,12 +36,12 @@ def _set_multiindex_dtype(index, level, type_):
 
 
 def _read_tsv(path_or_buffer) -> pd.DataFrame:
-    d = pd.read_csv(path_or_buffer, sep='\t', header=0, dtype=str)
+    d = pd.read_csv(path_or_buffer, sep="\t", header=0, dtype=str)
 
     top_left_cell = d.columns[0]
 
-    row_dimension_names, header_dimension_name = top_left_cell.split('\\')
-    row_dimension_names = row_dimension_names.split(',')
+    row_dimension_names, header_dimension_name = top_left_cell.split("\\")
+    row_dimension_names = row_dimension_names.split(",")
     index_data = d[top_left_cell]
     del d[top_left_cell]
     assert len(set(index_data)) == len(index_data)  # no duplicates
@@ -48,12 +50,11 @@ def _read_tsv(path_or_buffer) -> pd.DataFrame:
 
     d.columns.name = header_dimension_name
 
-    index_data = index_data.apply(lambda s: s.split(','))
+    index_data = index_data.apply(lambda s: s.split(","))
 
     d.index = pd.MultiIndex.from_arrays(
-        list(zip(*index_data)),
-        names=row_dimension_names,
-        )
+        list(zip(*index_data)), names=row_dimension_names,
+    )
 
     # cannot handle multidimensional column labels
     d = d.stack()
@@ -64,33 +65,32 @@ def _read_tsv(path_or_buffer) -> pd.DataFrame:
     assert all(map(_is_valid_dimension_name, d.index.names))
 
     d.index.set_levels(
-        [level.str.strip() for level in d.index.levels],
-        inplace=True
-        )
+        [level.str.strip() for level in d.index.levels], inplace=True
+    )
 
     d = _split_values_flags(d)
 
-    d.loc[d['value'] == ':', 'value'] = np.nan
-    d['value'] = d['value'].astype(float)
+    d.loc[d["value"] == ":", "value"] = np.nan
+    d["value"] = d["value"].astype(float)
 
-    if 'time' in d.index.names:
-        time_strings = d.index.unique('time')
+    if "time" in d.index.names:
+        time_strings = d.index.unique("time")
         matches_year = (_YEAR_RE.match(s) for s in time_strings)
         if all(matches_year):
-            d.index = _set_multiindex_dtype(d.index, 'time', int)
+            d.index = _set_multiindex_dtype(d.index, "time", int)
 
     d = d.sort_index()
 
     return d
 
 
-_TSV_GZ_FILENAME = 'data.tsv.gz'
-_HDF_FILENAME = 'data.h5'
-_HDF_TABLE_PATH = 'eurostat_table'
+_TSV_GZ_FILENAME = "data.tsv.gz"
+_HDF_FILENAME = "data.h5"
+_HDF_TABLE_PATH = "eurostat_table"
 
 
 def _read_tsv_gz(path_or_buffer) -> pd.DataFrame:
-    with gzip.open(path_or_buffer, 'rb') as f:
+    with gzip.open(path_or_buffer, "rb") as f:
         return _read_tsv(f)
 
 
@@ -110,9 +110,9 @@ def _read(the_dir):
         data.to_hdf(
             hdf_path,
             _HDF_TABLE_PATH,
-            complevel=conf['hdf_complevel'],
-            complib=conf['hdf_complib'],
-            )
+            complevel=conf["hdf_complevel"],
+            complib=conf["hdf_complib"],
+        )
 
     # Replace empty flags by None (issue #3)
     #
@@ -122,6 +122,6 @@ def _read(the_dir):
     # "your performance may suffer as PyTables will pickle object types
     # that it cannot map directly to c-types
     # [inferred_type->mixed,key->block0_values] [items->['flag']]"
-    data['flag'] = data['flag'].replace({'': None})
+    data["flag"] = data["flag"].replace({"": None})
 
     return data
